RobotInfoクラスをcrane_basicsへ移動
